Cleanup table code
